requirements.txt
# Derin Öğrenme ve Model
tensorflow==2.13.0
keras==2.13.1

# Görüntü İşleme
opencv-python==4.8.1.78
pillow==10.0.0

# Veri İşleme ve Analiz
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Görselleştirme
matplotlib==3.7.2
seaborn==0.12.2

# Utility
tqdm==4.66.1

# Not: tkinter Python ile birlikte gelir, ayrıca yüklemeye gerek yok
